restructure and fix empty report
